refactoring for more visibility.
