+++ examples/loading.py
@@ -1,0 +1,5 @@
+import spartacus as sp
+
+spartacus_dataset = sp.load()
+
+
